Refactoring into 'feature' directories
sorry for the big commit

- updates us to use react-query
- splits out API functionality across features
- adds locker list to homepage
- adds live input to zip/address input
