+++ frontend/package.json
@@ -13,10 +13,12 @@
     "flatlist-react": "^1.5.0",
     "google-maps-react": "^2.0.6",
     "googleplaces": "^0.6.1",
+    "moment": "^2.29.1",
     "node-fetch": "^2.6.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-map-gl": "^6.1.16",
+    "react-query": "^3.19.0",
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.3",
     "semver": "^7.3.5",
