+++ frontend/src/App.js
@@ -1,22 +1,26 @@
-import './App.css';
+import "./App.css";
 import React from "react";
 import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
 import { Navigation, Footer, Home, About, MyStuff } from "./components";
-import { NewAsk } from './screens'
+import { NewAsk } from "./screens";
+import { QueryClient, QueryClientProvider } from "react-query";
+const queryClient = new QueryClient();
 
 function App() {
   return (
     <div className="App">
-      <Router>
-        <Navigation />
-        <Switch>
-          <Route exact path="/" component={() => <Home />} />
-          <Route path="/ask" exact component={() => <NewAsk />} />
-          <Route path="/about" exact component={() => <About />} />
-          <Route path="/my-stuff" exact component={() => <MyStuff />} />
-        </Switch>
-        <Footer />
-      </Router>
+      <QueryClientProvider client={queryClient}>
+        <Router>
+          <Navigation />
+          <Switch>
+            <Route exact path="/" component={() => <Home />} />
+            <Route path="/ask" exact component={() => <NewAsk />} />
+            <Route path="/about" exact component={() => <About />} />
+            <Route path="/my-stuff" exact component={() => <MyStuff />} />
+          </Switch>
+          <Footer />
+        </Router>
+      </QueryClientProvider>
     </div>
   );
 }
