+++ frontend/src/features/askList/api.ts
@@ -1,0 +1,6 @@
+import { getLockerInfoFromPlaceId } from "../lockers/api";
+
+export const getInfoForLockers = async (lockerIds) => {
+  let results = lockerIds.map(getLockerInfoFromPlaceId);
+  return Promise.all(results);
+};
